# swagger.yaml
openapi: 3.0.0
info:
  title: Go Shop API
  version: 1.0.0
  description: API for a simple shop application.
servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /api/sign-up:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpInput'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpOutput'
        '400':
          description: Invalid input or user already exists

  /api/sign-in:
    post:
      tags:
        - Auth
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInInput'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInOutput'
        '401':
          description: Invalid credentials

  /api/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshInput'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshOutput'
        '401':
          description: Invalid refresh token

  /api/logout:
    post:
      tags:
        - Auth
      summary: Logout user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutInput'
      responses:
        '200':
          description: Logout successful
        '401':
          description: Invalid refresh token

  /api/logout-all:
    post:
      tags:
        - Auth
      summary: Logout from all devices
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized

  /api/list:
    post:
      tags:
        - Lists
      summary: Create a new list
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListInput'
      responses:
        '200':
          description: List created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
    get:
      tags:
        - Lists
      summary: Get all lists for a user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/List'
        '401':
          description: Unauthorized

  /api/list/{id}:
    get:
      tags:
        - Lists
      summary: Get a list by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '401':
          description: Unauthorized
        '404':
          description: List not found
    put:
      tags:
        - Lists
      summary: Update a list
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListInput'
      responses:
        '200':
          description: List updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: List not found
    delete:
      tags:
        - Lists
      summary: Delete a list
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List deleted successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        '401':
          description: Unauthorized
        '404':
          description: List not found

  /api/category:
    post:
      tags:
        - Categories
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateInput'
      responses:
        '200':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Wrong input
        '500':
          description: Failed to create category

  /api/category/{id}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
    put:
      tags:
        - Categories
      summary: Update category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateInput'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Wrong input
        '500':
          description: Failed to update category
    delete:
      tags:
        - Categories
      summary: Delete category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDeleteResponse'
        '404':
          description: Category not found

  /api/category/{id}/order:
    patch:
      tags:
        - Categories
      summary: Update category order within its level
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateOrderInput'
      responses:
        '200':
          description: Category order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Wrong input
        '500':
          description: Failed to update category order

  /api/category/{id}/status:
    patch:
      tags:
        - Categories
      summary: Update category active status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateStatusInput'
      responses:
        '200':
          description: Category status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Wrong input
        '500':
          description: Failed to update category status

  /api/category/tree:
    get:
      tags:
        - Categories
      summary: Get full category tree
      responses:
        '200':
          description: Category tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryTreeNode'
        '500':
          description: Failed to get category tree

  /api/category/flat:
    get:
      tags:
        - Categories
      summary: Get flat list of categories
      responses:
        '200':
          description: Flat category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

  /api/category/slug/{slug}:
    get:
      tags:
        - Categories
      summary: Get category by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category details by slug
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found

  /api/category/slug/{slug}/products:
    get:
      tags:
        - Categories
      summary: Get products for category by slug (including nested)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Products for category
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '404':
          description: Category not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignUpInput:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - username
        - password

    SignUpOutput:
      type: object
      properties:
        id:
          type: integer

    SignInInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    SignInOutput:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    RefreshInput:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken

    RefreshOutput:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    LogoutInput:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken

    List:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        user_id:
          type: integer

    ListInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title

    CategoryCreateInput:
      type: object
      properties:
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        displayOrder:
          type: integer
        isActive:
          type: boolean
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        metaKeywords:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        level:
          type: integer
        path:
          type: string
          nullable: true

    CategoryUpdateInput:
      type: object
      properties:
        parent_id:
          type: integer
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image_url:
          type: string
        display_order:
          type: integer
        is_active:
          type: boolean
        meta_title:
          type: string
        meta_description:
          type: string
        meta_keywords:
          type: string
        level:
          type: integer

    CategoryUpdateOrderInput:
      type: object
      properties:
        level:
          type: integer
      required:
        - level

    CategoryUpdateStatusInput:
      type: object
      properties:
        isActive:
          type: boolean
      required:
        - isActive

    Category:
      type: object
      properties:
        id:
          type: integer
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        displayOrder:
          type: integer
        isActive:
          type: boolean
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        metaKeywords:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        level:
          type: integer
        path:
          type: string
          nullable: true

    CategoryDeleteResponse:
      type: object
      properties:
        id:
          type: integer

    CategoryTreeNode:
      allOf:
        - $ref: '#/components/schemas/Category'
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: '#/components/schemas/CategoryTreeNode'